% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIPfilter.R
\name{CHIPfilter}
\alias{CHIPfilter}
\title{Extract potential CHIP from an annotated txt file with different filters.}
\usage{
CHIPfilter(
  input,
  max_percent = 0.1,
  max_percent_locus = 0.3,
  DP_min = 20,
  SOR_min = 3,
  Qual_min = 6.3,
  Alt_AD_min = 2,
  VAF_min = 0.02,
  VAF_max = 0.4,
  SAF_min = 0,
  SAR_min = 0,
  tumor_sample = F,
  tumor_VAF_min = 0.35,
  tumor_muti = 1,
  germline_VAF_min = 0.35,
  gnomad = T,
  gnomad_min = 0.005,
  dpGAP = F,
  dpGAP_min = 0.005,
  variant_effect_VAF_min = 0.25,
  not_nonsynonymous = T,
  blacklist_f = F,
  info = T
)
}
\arguments{
\item{input}{a data frame of text annotated file with proper column names.}

\item{max_percent}{maximum percentage of sample size allowed for one variant.}

\item{max_percent_locus}{maximum percentage of sample size allowed for one locus.}

\item{DP_min}{minimum read depth.}

\item{SOR_min}{minimum SOR value.}

\item{Qual_min}{minimum Qual(TLOD) value.}

\item{Alt_AD_min}{minimum ALT AD value.}

\item{VAF_min}{minimum VAF value.}

\item{VAF_max}{maximun VAF value.}

\item{SAF_min}{minimum SAF value.}

\item{SAR_min}{minimum SAR value.}

\item{tumor_sample}{if there are paired tumor samples.}

\item{tumor_VAF_min}{minimum VAF value in paired tumor sample. (Available when tumor_sample is used)}

\item{tumor_muti}{factor used for VAF comparison between blood and tumor samples. (Available when tumor_sample is used)}

\item{germline_VAF_min}{minimum VAF value in blood sample when there is paired tumor sample. (Available when tumor_sample is used)}

\item{gnomad}{if gnomad reference are used.}

\item{gnomad_min}{minimum non cancer AF popmax from gnomad reference file. (Available when gnomad is used)}

\item{dpGAP}{if dpGAP reference are used.}

\item{dpGAP_min}{minimum ALT dpGAP PopFreq value in blood sample.}

\item{variant_effect_VAF_min}{minimum VAF value in blood sample for variant effect filter. (Available when dpGAP or gnomad reference are used)}

\item{not_nonsynonymous}{if including not nonsynonymous option.}

\item{blacklist_f}{a bed file contain blacklist regions to exclude.}

\item{info}{if print processing massage.}
}
\value{
A text file with CHIP candidates.
}
\description{
A function to extract CHIP with different filters
}
\details{
This function will return a txt file of CHIP candidates based on the filter settings.
}
\examples{
# input file
in_df <- fread("data/demo_input.txt")
# run default setting
out_1 <- CHIPfilter(in_df)

# change different metrics
out_2 <- CHIPfilter(in_df,max_percent=0.02,DP_min = 40,VAF_min=0.002)

# with paired tumor sample
out_3 <- CHIPfilter(in_df,tumor_sample = T,tumor_VAF_min = 0.02)

# with gnomad or dpGAP reference file
out_4 <- CHIPfilter(in_df,gnomad = F,dpGAP = T)

# with blacklist region
bl_f <- fread("blacklist.bed")
out_5 <- CHIPfilter(in_df,blacklist_f = bl_f)
}
