% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCHIP.R
\name{qcCHIP}
\alias{qcCHIP}
\title{Guide the selection of VAF, DP, or population metric through permutation analysis}
\usage{
qcCHIP(
  input_df,
  blacklist_f = F,
  out_path,
  permut_metrics = "VAF",
  permut_n = 20,
  sub_sample_n = c(2, 5, 10),
  metric_min = 0,
  metric_max = 0.1,
  metric_step = 0.05,
  remove_tmp_dir = T,
  core = 2,
  show_info = T
)
}
\arguments{
\item{input_df}{a data frame of text annotated file with proper column names. (detailed in instruction)}

\item{blacklist_f}{a bed file contain blacklist regions to exclude.}

\item{out_path}{path of output directory}

\item{permut_metrics}{which metric want to permute.
VAF: minimum VAF setting;
DP: minimum read depth;
population: maximum percentage of sample size allowed for one variant.}

\item{permut_n}{number of permutations.}

\item{sub_sample_n}{integers of subset sample groups.
For example: c(2,5) means the function will split the whole sample into 2 and 5 groups, each group has same size.}

\item{metric_min}{minimum value to start with.}

\item{metric_max}{maximum value to end with.}

\item{metric_step}{the step of metric changes.}

\item{remove_tmp_dir}{if remove temporary directory.}

\item{core}{number of cores used for parallel computing.}

\item{show_info}{if print processing massage.}
}
\value{
A text file with comparison results of subsample and whole sample will be saved and written to out_path.
A pdf file including 3 permutation consistency plots will be written to out_path.
A temp directory contains all permutation and whole sample output files of metric setting for all sub-sample approaches (if remove_tmp_dir==F).
}
\description{
A function of generating plots of permutation consistency based on different setting of VAF, DP, or population metric.
}
\details{
This function will split the whole input file to N groups.
A permutation will be used to compare the results from subseting sample and whole sample.
Finally, a summary file and a permutation consistency plot will be generated to guide the selection of VAF, DP, or population metric.
}
\examples{
# input file
in_df <- fread("data/demo_input.txt")
bl_f <- fread("blacklist.bed")

# permutation on VAF
qcCHIP(input_df,
      out_path = "out_path",
      metric_min = 0,
      metric_step = 0.005,
      metric_max=0.1,
      permut_metrics = "VAF")

# permutation on DP with blacklist region
qcCHIP(input_df,
      out_path = "out_path",
      metric_min = 0,
      metric_step = 5,
      metric_max=50,
      permut_metrics = "DP",
      blacklist_f = bl_f)

# permutation on population with blacklist region and do not print info
qcCHIP(input_df,
      out_path = "out_path",
      metric_min = 0.05,
      metric_step = 0.005,
      metric_max=0.2,
      permut_metrics = "population",
      blacklist_f = bl_f,
      show_info=F)
}
