% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2input.R
\name{vcf2input}
\alias{vcf2input}
\title{Convert individual VCF file to qcCHIP input table.}
\usage{
vcf2input(
  vcf_path,
  sample_list,
  tumor = F,
  tumor_path,
  tumor_list,
  refGene = F,
  refGene_func_name = NA,
  refGene_gene_name = NA,
  refGene_Exonicfunc_name = NA,
  refGene_AAchange_name = NA,
  refGene_GeneDetail_name = NA,
  cosmic = F,
  cosmic_name = NA,
  gnomad = F,
  gnomad_name = NA,
  dpGAP = F,
  dpGAP_name = NA
)
}
\arguments{
\item{vcf_path}{Path of VCF files of individual samples.}

\item{sample_list}{VCF file names of individual samples.}

\item{tumor}{If there are paired tumor samples.}

\item{tumor_path}{Path of VCF files of paired tumor samples.}

\item{tumor_list}{VCF file names of paired tumor samples.}

\item{refGene}{If there is annotation INFO of refGene.}

\item{refGene_func_name}{Name that contains the value function of gene (e.g. exonic, UTR3, ncRNA_intronic, etc.).}

\item{refGene_gene_name}{Name that contains the value name of gene.}

\item{refGene_Exonicfunc_name}{Name that contains the value exonic function of gene (e.g. stopgain, synonymous_SNV, nonsynonymous_SNV, etc.).}

\item{refGene_AAchange_name}{Name that contains the value AAchange of gene.}

\item{refGene_GeneDetail_name}{Name that contains the value Genedetail of gene.}

\item{cosmic}{If there is annotation from cosmic database.}

\item{cosmic_name}{Name that contains the cosmic value.}

\item{gnomad}{If there is annotation from gnomad database.}

\item{gnomad_name}{Name that contains the gnomad value.}

\item{dpGAP}{If there is annotation from dpGAP database.}

\item{dpGAP_name}{Name that contains the dpGAP value.}
}
\value{
a table for qcCHIP
}
\description{
A function to convert individual VCF file to table for qcCHIP
}
\details{
This function will convert individual VCF file to qcCHIP input file.
}
\examples{

# No paired tumor sample and no annotation of other database
out_1 <- vcf2input(vcf_path = vcf_path,
                   sample_list=sample_list)

# With paired tumor sample and no annotation of other database
out_2 <- vcf2input(vcf_path = vcf_path,
                   sample_list=sample_list,
                   tumor=T,
                   tumor_path=tumor_path,
                   tumor_list=tumor_list)

# Annotated VCF of sample and paired tumor sample. Annotated with refGene, cosmic, and gnomad.
out_3 <- vcf2input(vcf_path = vcf_path_annot,
                   sample_list=sample_list_annot,
                   tumor=T,
                   tumor_path=tumor_path_annot,
                   tumor_list=tumor_list_annot,
                   refGene=T,
                   refGene_func_name="Func.refGeneWithVer",
                   refGene_gene_name="Gene.refGeneWithVer",
                   refGene_Exonicfunc_name="ExonicFunc.refGeneWithVer",
                   refGene_AAchange_name=NA,
                   refGene_GeneDetail_name=NA,
                   cosmic=T,
                   cosmic_name="cosmic70",
                   gnomad=T,
                   gnomad_name="non_cancer_AF_popmax")
}
